[FIX] 게이트 웨이 필터링 문제 해결 및 CORS 설정

+++ common/src/main/java/com/jh/common/utils/ParseRequestUtil.java
@@ -4,13 +4,13 @@ import jakarta.servlet.http.HttpServletRequest;
 
 public class ParseRequestUtil {
 
-    public int extractUserIdFromRequest(HttpServletRequest request) {
+    public Long extractUserIdFromRequest(HttpServletRequest request) {
         String userIdHeader = request.getHeader("x-claim-userid");
         if (userIdHeader == null) {
             throw new RuntimeException("헤더에 사용자 ID가 없습니다.");
         }
 
-        int userId;
+        long userId;
         try {
             userId = Integer.parseInt(userIdHeader);
         } catch (NumberFormatException e) {
